+++ src/game/SpellAuras.h
@@ -551,7 +551,7 @@ public:
 	//void SpellAuraModRangedDamageTakenPCT(bool apply);
 	void SpellAuraModBlockValue(bool apply);
 
-	void SendModifierLog(int32 ** m,int32 v,uint32 mask,uint8 type,bool pct = false);
+	void SendModifierLog(int32 ** m,int32 v,uint64 mask,uint8 type,bool pct = false);
 	void SendDummyModifierLog(std::map<SpellEntry*,uint32> * m,SpellEntry * spellInfo,uint32 i,bool apply,bool pct = false);
 
 	// Events
